Add column for item no and sync from bus (#165)
* add column for ItemNo

* extract contant, removed unused import

* Make ItemNo unique

+++ src/Equinor.ProCoSys.Completion.Domain/Events/IntegrationEvents/PunchItemEvents/PunchItemCreatedIntegrationEvent.cs
@@ -15,7 +15,7 @@ public record PunchItemCreatedIntegrationEvent
     string ProjectDescription,
     Guid CheckListGuid,
     string Category,
-    int ItemNo,
+    long ItemNo,
     string Description,
     string RaisedByOrgCode,
     Guid RaisedByOrgGuid,
