+++ src/tests/Equinor.ProCoSys.Completion.MessageContracts.Tests/PunchItem/PunchItemUpdatedV1ContractTests.cs
@@ -21,7 +21,7 @@ public class PunchItemUpdatedV1ContractTests : ContractTestBase<IPunchItemUpdate
             { "ProjectName", typeof(string) },
             { "ProjectDescription", typeof(string) },
             { "Category", typeof(string) },
-            { "ItemNo", typeof(int) },
+            { "ItemNo", typeof(long) },
             { "Description", typeof(string) },
             { "RaisedByOrgCode", typeof(string) },
             { "RaisedByOrgGuid", typeof(Guid) },
